scripts/story-wireframes/requirements.txt
openai>=1.0.0
PyYAML>=6.0
python-slugify>=8.0.0
sumy>=0.11.0
nltk>=3.8
requests>=2.31.0
tqdm>=4.66.0
pytest>=7.4.0
